update shared data

+++ src/tests/test_trait.cpp
@@ -44,7 +44,7 @@ int main(int argc, char *argv[]) {
     t.f = make_float3x3(56.1f);
     ResourceArray ra = device.create_resource_array();
 
-    ManagedWrapper<float> vv(ra);
+    ManagedWrapper<int> vv(ra);
 
     t.encode(vv);
     vv.reset_device_buffer(device);
@@ -56,7 +56,7 @@ int main(int argc, char *argv[]) {
 
 
     Kernel kernel = [&](Float a) {
-        DataAccessor<float> da{0u, vv};
+        DataAccessor<int> da{0u, vv};
         t.decode(&da);
         Printer::instance().info("a = {} {}", t.a.dv());
         Printer::instance().info("b = {} {} {}", t.b.dv());
